allow a scope owner to request data for their own scope
