+++ server/src/main/kotlin/io/provenance/objectstore/gateway/server/ObjectStoreGatewayServer.kt
@@ -21,10 +21,7 @@ import java.time.OffsetDateTime
 
 @GRpcService()
 class ObjectStoreGatewayServer(
-    private val scopePermissionsRepository: ScopePermissionsRepository,
     private val scopeFetchService: ScopeFetchService,
-    private val provenanceProperties: ProvenanceProperties,
-    private val encryptionKeys: Map<String, KeyRef>
 ): GatewayGrpc.GatewayImplBase() {
 
     companion object : KLogging()
@@ -33,8 +30,6 @@ class ObjectStoreGatewayServer(
         request: GatewayOuterClass.FetchObjectRequest,
         responseObserver: StreamObserver<GatewayOuterClass.FetchObjectResponse>
     ) {
-        val requesterPublicKey = request.signature.signer.signingPublicKey.toPublicKey()
-
         if (request.params.expiration.isNotSet() || request.params.expiration.toOffsetDateTime().isBefore(OffsetDateTime.now())) {
             logger.info("Request with invalid expiration received")
             throw TimestampValidationException("${if (request.params.expiration.isNotSet()) "Missing" else "Expired"} expiration")
@@ -45,21 +40,7 @@ class ObjectStoreGatewayServer(
             throw SignatureValidationException("Invalid signature")
         }
 
-        val requesterAddress = requesterPublicKey.getAddress(provenanceProperties.mainNet)
-        val granterAddress = scopePermissionsRepository.getAccessGranterAddress(request.params.scopeAddress, requesterAddress, request.params.granterAddress.takeIf { it.isNotBlank() })
-        if (granterAddress == null) {
-            logger.info("Request for scope data without access received")
-            throw AccessDeniedException("Scope access not granted to $requesterAddress")
-        }
-
-        val encryptionKey = encryptionKeys[granterAddress]
-        if (encryptionKey == null) {
-            logger.warn("Valid request for scope data with an unknown key received")
-            throw AccessDeniedException("Encryption key for granter $granterAddress not found")
-        }
-
-        logger.debug("Valid request for scope data received")
-        scopeFetchService.fetchScope(request.params.scopeAddress, encryptionKey).let {
+        scopeFetchService.fetchScope(request.params.scopeAddress, request.signature.signer.signingPublicKey.toPublicKey(), request.params.granterAddress.takeIf { it.isNotBlank() }).let {
             responseObserver.onNext(GatewayOuterClass.FetchObjectResponse.newBuilder()
                 .setScopeId(request.params.scopeAddress)
                 .addAllRecords(it)
